ProjectUTS/requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
django-cors-headers==4.3.1
sqlparse==0.5.3
tzdata==2025.2
PyJWT==2.8.0
